bot/requirements.txt
ncatbot>=0.1.0
volcenginesdkarkruntime>=0.1.0
aiohttp>=3.8.0
pyyaml>=6.0